Refactor: Avoid circular imports
